Add swatches to outline menu
