+++ src/containers/stroke-color-indicator.jsx
@@ -6,6 +6,7 @@ import {changeStrokeColor} from '../reducers/stroke-color';
 import {changeStrokeWidth} from '../reducers/stroke-width';
 import {openStrokeColor, closeStrokeColor} from '../reducers/modals';
 import Modes from '../lib/modes';
+import getAllColors from '../lib/colors';
 import Formats from '../lib/format';
 import {isBitmap} from '../lib/format';
 
@@ -54,6 +55,7 @@ class StrokeColorIndicator extends React.Component {
         return (
             <StrokeColorIndicatorComponent
                 {...this.props}
+                colors={getAllColors()}
                 onChangeStrokeColor={this.handleChangeStrokeColor}
                 onCloseStrokeColor={this.handleCloseStrokeColor}
             />
